Move auth pkg protocol to root lib.rs
