+++ windows_integration/public/src/lib.rs
@@ -1,1 +1,20 @@
-pub mod secpkg;
+pub enum AccountType {
+	Person,
+	Service,
+}
+
+// Authentication Package Protocol
+pub enum AuthPkgRequest {
+	AuthenticateAccount(AuthenticateAccountRequest),
+}
+
+pub enum AuthPkgResponse {
+	AuthenticateAccount(AuthenticateAccountResponse),
+	Error(AuthPkgError),
+}
+
+pub enum AuthPkgError {}
+
+pub struct AuthenticateAccountRequest {}
+
+pub struct AuthenticateAccountResponse {}
