--- windows_integration/public/src/secpkg/ap_proto/mod.rs
@@ -1,9 +1,0 @@
-pub mod v1;
-
-pub enum AuthPkgRequest {
-	V1(v1::AuthPkgRequest),
-}
-
-pub enum AuthPkgResponse {
-	V1(v1::AuthPkgResponse),
-}
