--- windows_integration/public/src/secpkg/mod.rs
@@ -1,1 +1,0 @@
-pub mod ap_proto;
